+++ source/Private/Remove-MarkdownMetadata.ps1
@@ -3,7 +3,7 @@
         Removes metadata from a Markdown file.
 
     .DESCRIPTION
-        The Remove-MarkdownMetadata function removes metadata from a Markdown file.
+        The Remove-MarkdownMetadataBlock function removes metadata from a Markdown file.
         It searches for a metadata marker ('---') and removes the content between
         the marker and the next occurrence of the marker.
 
@@ -11,12 +11,12 @@
         Specifies the path to the Markdown file from which the metadata should be removed.
 
     .EXAMPLE
-        Remove-MarkdownMetadata -FilePath 'C:\Path\To\File.md'
+        Remove-MarkdownMetadataBlock -FilePath 'C:\Path\To\File.md'
 
         Removes the metadata from the specified Markdown file.
 #>
 
-function Remove-MarkdownMetadata
+function Remove-MarkdownMetadataBlock
 {
     [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSUseShouldProcessForStateChangingFunctions', '', Justification = 'This function is a private helper function and is not exported publicly.')]
     [CmdletBinding()]